Move handelsregister registrations to own class
